Refactoring: Extract StringExtensions
